Use asmjit "kInstOptionShortForm" option for instructions
Without it, immediates to instructions (e.g., jump offsets) are always encoded
as 32-bit constants which may make instructions larger than they will be when
a standalone assembler encodes them, changing the alignment of the following
instructions. This may affect run-time behaviour, such as which instructions
get decoded in a given clock cycle, if instructions cross the fetch window.