If a line contains "ajs:notshortform" after a comment character, don't use asmjit kInstOptionShortForm for that instruction
Displacements in non-trivial functions may not fit into a signed 8-bit offset,
so we need to be able to encode those with a larger offset. This patch allows
annotating lines in the input file to tell asmjit not to use the
kInstOptionShortForm option for that line. Unfortunately, this involves
manual work to find out which lines require this annotation. Running
ajs with -vv helps here, as the asmjit assembler emits error messages for
offsets that can't be encoded with kInstOptionShortForm, and the source
line that is printed with each line that is passed to asmjit makes finding
the corresponding line in the input file reasonably easy.