Pad stack pointer to a fixed address modulo 4096
This greatly reduces the variability of timings between different ajs runs.

If a memory read goes to an address that overlaps (modulo 4096) with a
recently-written-to address, the CPU has to do full
virtual-to-physical address translation to determine whether there is
a RAW-dependency. This costs several clock cycles and can be measured
with the ld_blocks_partial.address_alias profiling event.

Depending on what the value of the stack pointer is, the function
entry (CALL, ENTER, PUSHes) may write to addresses that overlap
(modulo 4096) with input data which the function reads shortly after
function entry, causing the aforementioned penalties.